Active Storage: remove transformers
They were introduced to support the transition from MiniMagick to ImageProcessing. We're fully on ImageProcessing now, so they're no longer needed. Use ImageProcessing directly from Variation.
