+++ activestorage/lib/active_storage.rb
@@ -73,11 +73,4 @@ module ActiveStorage
   mattr_accessor :video_preview_arguments, default: "-y -vframes 1 -f image2"
 
   mattr_accessor :silence_invalid_content_types_warning, default: false
-
-  module Transformers
-    extend ActiveSupport::Autoload
-
-    autoload :Transformer
-    autoload :ImageProcessingTransformer
-  end
 end
